+++ http/ruby/nethttp/microsoft_kiota_nethttplibrary/spec/microsoft_kiota_nethttp_spec.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+require_relative 'authentication_provider'
 
 RSpec.describe MicrosoftKiotaNethttplibrary do
   it "has a version number" do
@@ -7,5 +8,11 @@ RSpec.describe MicrosoftKiotaNethttplibrary do
 
   it "does something useful" do
     expect(true).to eq(true)
+  end
+
+  it "can instantiate a request adapter" do
+    auth_provider = AuthenticationProvider.new()
+    adapter = MicrosoftKiotaNethttplibrary::NetHttpRequestAdapter.new(auth_provider)
+    expect(adapter).to_not be nil
   end
 end